Convert to TS

+++ final/server/src/datasources/user.ts
@@ -1,10 +1,16 @@
 import { DataSource } from "apollo-datasource";
 import * as isEmail from "isemail";
+import { Context, Store } from "..";
+import { StoreTrip, SQLMany, SQLOne, StoreUser } from "../utils";
+
+function nonNill<T>(x: T | undefined): x is T {
+  return x !== undefined;
+}
 
 export class UserAPI extends DataSource {
-  private context: any;
-  private store: any;
-  constructor({ store }) {
+  private context: Context | undefined;
+  private store: Store;
+  constructor({ store }: { store: Store }) {
     super();
     this.store = store;
   }
@@ -15,7 +21,7 @@ export class UserAPI extends DataSource {
    * like caches and context. We'll assign this.context to the request context
    * here, so we can know about the user making requests
    */
-  initialize(config) {
+  initialize(config: any) {
     this.context = config.context;
   }
 
@@ -33,12 +39,14 @@ export class UserAPI extends DataSource {
       this.context && this.context.user ? this.context.user.email : emailArg;
     if (!email || !isEmail.validate(email)) return null;
 
-    const users = await this.store.users.findOrCreate({ where: { email } });
+    const users = (await this.store.users.findOrCreate({
+      where: { email }
+    })) as [StoreUser, boolean];
     return users && users[0] ? users[0] : null;
   }
 
   async bookTrips({ launchIds }) {
-    const userId = this.context.user.id;
+    const userId = this.context && this.context.user.id;
     if (!userId) return [];
 
     let results: any[] = [];
@@ -54,25 +62,28 @@ export class UserAPI extends DataSource {
   }
 
   async bookTrip({ launchId }) {
+    if (!this.context) return false;
     const userId = this.context.user.id;
-    const res = await this.store.trips.findOrCreate({
+    const res = (await this.store.trips.findOrCreate({
       where: { userId, launchId }
-    });
+    })) as [SQLOne<StoreTrip>, boolean];
     return res && res.length ? res[0].get() : false;
   }
 
   async cancelTrip({ launchId }) {
+    if (!this.context) return false;
     const userId = this.context.user.id;
     return !!this.store.trips.destroy({ where: { userId, launchId } });
   }
 
   async getLaunchIdsByUser() {
+    if (!this.context) return false;
     const userId = this.context.user.id;
-    const found = await this.store.trips.findAll({
+    const found = (await this.store.trips.findAll({
       where: { userId }
-    });
+    })) as SQLMany<StoreTrip>;
     return found && found.length
-      ? found.map(l => l.dataValues.launchId).filter(l => !!l)
+      ? found.map(l => l.dataValues.launchId).filter(nonNill)
       : [];
   }
